requirements.txt
Pillow==10.4.0
torch==2.4.0
torchvision==0.19.0
